requirements.txt
PyGithub==1.25.2
web.py==0.37
